Sorbus: add handshake to block read/write routines for internal drive
The Sorbus Computer is planned to get a change that requires handshaking
when a block is read or written from/to internal storage. While it is not
required now, the mechanism is already in place since the current release.
A good time to add it here, since it is very likely that this will be
required in the future.

Signed-off-by: Sven Oliver Moll <svolli@svolli.de>